add test to rlp tx test
